refactor: viewportHeight Hook extraction
