+++ components/chat/ChatList.jsx
@@ -1,8 +1,8 @@
-import ChatListItem from '@/components/chat/ChatListItem.jsx';
+import ChatListItem from '@/components/chat/ChatListItem.jsx'
 
 const ChatList = () => {
   return (
-    <div className='h-96 overflow-y-auto rounded-lg bg-white p-4 shadow-inner'>
+    <div className="mb-4 h-full overflow-y-auto rounded-lg bg-white p-4 shadow-inner scrollbar-hide">
       <ChatListItem />
       <ChatListItem />
       <ChatListItem />
@@ -10,9 +10,17 @@ const ChatList = () => {
       <ChatListItem />
       <ChatListItem />
       <ChatListItem />
+      <ChatListItem />
+      <ChatListItem />
+      <ChatListItem />
+      <ChatListItem />
+      <ChatListItem />
+      <ChatListItem />
+      <ChatListItem />
+
       {/* Add more ChatListItem components as needed */}
     </div>
-  );
-};
+  )
+}
 
-export default ChatList;
+export default ChatList
